Rework solve{Lower,Upper}Triangular* methods
We currently have to solve the following systems:
 - LU factorization (with L unitriangular):
    L*X = B
    U*X = B
    X*U = B
 - LDLt factorization (with L unitriangular):
    L*X = B
    D*X = B
    t(L)*X = B
    X*t(L) = B

solveUpperTriangular{,Left} has an extra argument to disambiguate
U*X and t(L)*X; in the latter case, L is unitriangular, otherwise
U is not.

But we want to add the following methods for LLt factorization
(with L non unitriangular):
    L*X = B
    t(L)*X = B
    X*t(L) = B

We could decide to store factorization method into HMatrix structure,
but this will make changes much more difficult when adding new factorization
methods.
Instead we add a boolean argument to tell whether triangular matrix is
unitriangular or not.

In order to ensure that all calls to these solve* functions have been checked,
arguments are no more optional.