A new RkApproximationControl member in RkMatrix Class.
This member replaces CompressionMethod.

With this member we control local compression/truncate  (in a HMatrix node).

The choice of RkApproximationControl is done in the  HMatrix::assemble function.

All Rk algebric operations use a specific RkApproximationControl
inherited from the assemble step.

All compression methods use a specific RkApproximationControl different from
the global one.

Note :
Combined methods in RkMatrix<T>::formattedAddParts is not necessary but
it is better to choose the RkApproximationControl with the highest epsilon!