New method to truncate the matrix after factorization.
This method is used to control the ranking of the factored matrix.
The function of truncation is applied recursively and effectively
on all Rk matrix.

The user has the possibility to build its own function truncation, by using
"truncateGeneric" function.

Or use the truncateEpsilon with a new value of epsilon recompression.
The default value of epsilon is the approx.recompressionEpsilon!

The epsilon can't be less then approx.recompressionEpsilon!