Fix AcaPlus compression with pivot = 0
When the pivot = 0 the new vector is zero and therefore it is not
necessary to renormalize it.

This situation can happen because the  shutoff is an approximate formula.

The disadvantage of this correction is that a null vector
will be added to "a" or "b".

Note that "recompression"  eliminates null vectors.