Allow the user to assemble directly the Rk matrix.
After preparing the block we know that this block will
be directly assembled in compressed form or not.
If is_user_compress is true then the classical compute rows
and cols will provide directly a and b.

is_null_col must be implemanted at the same time to calculate
the rank k (k <=min(rows, cols)).
The compute column i function must return column "i" of "a"
and the compute row i function must return the associated
column "i" in "b".