Add flag is_ortho to ScalarArray (WIP: could impact the performance)
The flag is stored in an int*, because it is associated to the buffer T*
and when the buffer address is copied (to create a subset), the flag
address is copied too. We have a flag ownOrtho to set if we own
'is_ortho' (and we can free it), it is different from ownMemory in the
cases where a constructor uses a T* as input.
We have accessor setOrtho() and getOrtho().
setOrtho(0) is called everytime we access to the buffer in RW mode,
in the accessor get() and ptr(). So it can have an impact in performance
for the routine that uses these functions a lot...