Add new parameter 'initialPivot' in MGStruncate()
initialPivot is the number of orthogonal columns in panels a and b.
It is transmitted to modifiedGramSchmidt().
If some columns [0..j] are already orthogonal in A, it can be
interesting to use these as pivots (and spare orthogonalisation within
these columns). The optionnal parameter initialPivot indicates the
number of columns [0..initialPivot-1] orthogonal.
It is used when called from formattedAddParts().
Deactivated by default, it is activated by setting HMAT_MGS_INITPIV.