modifiedGramSchmidt(): add option to use blas3 in classical gram-schmidt algo
We use blas3 (more computationnaly efficient) to apply classical
Gram-Schmidt approach to the 'initialPivot' first columns. With this
approach, since we do a classical (not a "modified") Gram Schmidt
algorithm which is known to be numerically stable, we expect to lose
several digits on final accuracy of hmat.
Activate with env. var. HMAT_MGS_BLAS3