Add new optionnal parameter 'initialPivot' in truncate()
Similar to what was done in truncateMGS().
initialPivot is the number of orthogonal columns in panels a and b.
It is transmitted to qrDecomposition(). default value is 0.
if initialPivot>0, on output of qrDecomposition() the 'initialPivot'
first columns of 'this' are directly the orthonormalized columns of Q,
and the remaining columns are produced by lapack ZGETRF (using
Householder reflectors).
Hence, when computing a product by Q, we must first call ZUNMQR via
productQ() for the columns >= initialPivot, and then GEMM for the columns
< initialPivot.
Deactivated by default: initialPivot set to 0 when entering truncate().
Activate with env. var. HMAT_TRUNC_INITPIV.