formattedAddParts(): improve the algorithm to reorder the Rk matrices
We try to find if the QR factorisation can be accelerated using
orthogonality information. The 1st possible optimization is to put in
first position the Rk matrix with orthogonal panels AND maximum rank.
Roughly, the gain from an initial pivot 'p' in a QR factorisation 'm x n'
is to reduce the flops count from 2mn^2 to 2m(n^2-p^2), so the gain grows
 like p^2 for each panel hence the gain formula to find the best
candidate: number of orthogonal panels x rank^2.
The second possible optimization is that when coallescing Rk from childs
toward parent, it is possible to "merge" Rk from (extra-)diagonal childs
because with non-intersecting rows and cols we will extend orthogonality
between separate Rk. So in this case we try to find the pair of Rk to put
in first positions that maximises the gain.