Change the way singular values are truncated
The former criteria was: sigma [k] / SUM (sigma) <epsilon
The new criteria is: sigma [k] / sigma[0] <epsilon
The previous criteria can be obtained by setting env. var.
	HMAT_SUM_CRITERION=1

The new criteria yields constant accuracy when the size of the test
cases increases and better accuracy, but of course computation is slower.