Make recursiveGemm more robust
This implementation works even if a, b and c have been
split by their admissibility conditions in a single direction.

When computing C <- C + A * B with A, B and C being split into
blocks, we have C_ij += A_ik * B_kj when dimensions of blocks
match.
But if we consider more exotic splits, as is done with RHS for
instance, this becomes more tricky.
We can write
  C_ij += sum(A_mk * B_nl)
for all k, l, m, n, by considering only overlaps of degrees
of freedom.