Fix cache cleaning in clustering algorithms
Method clean() was not called at the end of clustering, and thus
AdmissibilityCondition::prepare() did nothing.  But this caused crashes
because clustering algorithms do not create bounding boxes on leaves.

We could create bounding boxes on leaves, but it is safer to ensure that
clustering algorithms remove their cache themselves, admissibility
conditions may store objects which are not AxisAlignedBoundBox.