Descend clusterTrees by checking superset rather than depth
It was made for when the corresponding cluster trees are the same,
which is mathematically what we want, but too restrictive.
e.g. what if there are some "level0" shenanigans in the cluster tree ?

WIth this commit:

If the clustertrees are the same, it's the same as checking cluster depth.

In more general cases, it does something not dumb rather than looking at
a non-information.

TODO : what would be a better condition to descend in the general case ?
- Being a strict superset ?
- Having some spillage without necessarily containing all of the
  (allegedly) smaller matrix ?
- Checking the children rather than the parent (goes with the intersect check) ?