Fix recursiveSolve* recursions
The three solve[Lower/Upper]Triangular[Right/Left]
handle the recursion stopping condition (this->isLeaf or b->isLeaf)
while recursiveSolve* handle digging the matrices
(and expects to be able to do so)

Before this commit, one situation would try to apply recursiveSolve* twice
(or more) in a row without going back to solve*, this is a problem
if the leaves are reached after the first recursiveSolve.

In the case of solveLowerTriangularLeft, this is the following situation :

L->RecursiveSolve(b) is searching x11 x12 so that

[ L11 |  0  ]    [     |     ]   [     |     ]
[ ----+---- ] *  [ X11 | X12 ] = [ b11 | b12 ]
[ L21 | U22 ]    [     |     ]   [     |     ]

will call recursiveSolve on :
L->recursiveSolve(b11)

[ L11 |  0  ]    [     ]   [     ]
[ ----+---- ] *  [ X11 ] = [ b11 ]
[ L21 | U22 ]    [     ]   [     ]

When b11 is a leaf, it is not possible to dig any matrix and recursiveSolve
crashes.
With this commit, solve is called instead, stops the recursion and
finalize the operation.

This souldn't change anything to cases that happily called recursiveSolve*
twice in a row, as the solve* will immediatly call recursiveSolve* with
the same arguments after checking that the matrices are not leaves.