WIP Force children of blocs splitting both row and cols to do the same
Splitting only rows or only cols should only be done on rectangular matrices
to make the children bloc more square and thus easier to compress.

This is managed by the ratio setting of an admissibilty and usually,
a square bloc split in four will make square children. A ratio of sqrt(2)
should ensure existence and unicity of what is a square bloc.

However, some rounding happens when splitting an odd number of rows
that will break the inheritance of the aspect ratio. if the aspect ratio of a
matrix is close to sqrt(2), we may get "spinning" rectanguilar splits, when
rows/cols alternate between sqrt(2) and 1/sqrt(2).

This is problematic has it can be a cause of uncompatible cluster trees
during operations, when we can't match the cluster trees of all involved
matrix (see uncompatibleGemm).

This commit make spliting both rows and cols inheritable, thus performing all
mono-cluster splits at the start of the tree, and all bi-cluster splits at the
bottom. Limiting the risk of causing uncompatibities to the higher levels, and
allowing the uncompatibilities to be resolved before reaching the leaves.