Split admissibility ratio_ into a target rows/cols and a tolerance ratio
Assuming a cluster created via bissection, using a ratio_ of sqrt(0.5)
should be the reasonnable choice, giving existence of unicity of what
is the "squarest" a block can be according to splitRowsCols,
which is the interval ]sqrt(0.5, sqrt(2)[.

However, splitting odd sized clusters, or other rules that create unbalanced
cluster children sizes can be dangerous if a block rows/cols is close to the
borns of this intervals,

e.g. splitting a (29, 10) block into (15, 10) and (14, 10) produces one "skinny"
block split in 2 (8, 10) and (7, 10) and one "square", split in four blocks
(7, 5). This can cause blocks uncompatibilities errors.

To prevent this, we add targetRC, which should usually be set to approximately
the expected best ratio rows/cols a block can be, which the element of
{(row/2^n) / (cols/2^m), n, m in N} closest to 1, with rows and cols the
sizes at the root of the matrix.

ratio_ thus becomes a tolerance parameters, I suggest keeping it to sqrt(1/n)
where n is the number of children of the clusters.