Implement wrappers toward CUBLAS for geqrf
Based on the model of proxy_lapack::*, we add in cuda_manager.hpp
a namespace called proxy_cuda with wrappers toward cublas and cusolver
routines.
It implements geqrf in all 4 arithmetics of hmat, and it is used in
RkMatrix::truncate() like this:

T *tauA_gpu = nullptr, *Ra_gpu = nullptr;
info = proxy_cuda::geqrf(a->rows, a->cols, a_gpu, a->rows,
   &tauA_gpu, &Ra_gpu);

Technically, it is a bit more than just geqrf, since it also extracts
the R factor from A and copies it in a separate array.