Change recursiveGemm to keep blocs mostly compatible
It is not generally possible to keep blocs compatible through the recursion
without imposing strong contraints on the matrices structures.

With this, we keep the clusters the clusters of C equals to
their equivalent in A and B (A.rows and B.cols if no transposition).
so that uncompatibleGemm only needs to truncate one read-only matrix/cluster,
(either A.cols or B.rows) and not C.

RMQ : another induction constant with the same aim would be
"colA==rowB, rowA>=rowC and colB>=colC"

Also split recursiveGemm and computeGemmRecursion so that recursiveGemm
only handles the actual recursion, and any case where we do not actually
dig the matrices (e.g. uncompatibleGemm) is called directly by gemm.