Change recursiveGemm to keep blocs mostly compatible
It is not generally possible to keep blocs compatible through the recursion
without imposing strong contraints on the matrices structures.

With this, we aim to keep the clusters of A and B equals while C is allowed
to be smaller (but the least deep  as possible).

This is so that unCompatibleGemm/makeCompatible needs only to read sub-blocks
when A and B are too big and not write sub-blocks in a big C.

RMQ : it looks like "conditionX" and "childrenX" were the same, this commit
merge and rename them into "dig_x", which is more explicit on what's happening.

RMQ2 : another induction constant with the same aim would be
"rowA==rowC, colB==colC with A.depth and B.depth minimal"